Split ExecutionValidator
